+++ website/modules/handler.php
@@ -1,4 +1,29 @@
 <?php 
+// INPUT VALIDATION ERROR
+class regexException extends Exception {
+    public function errorMessage() {
+        return "Error ({$this->getFile()}:{$this->getLine()}) -> {$this->getMessage()}";
+    }
+}
+// DATABASE CONNECTION ERROR
+class databaseConnectException extends Exception {
+    public function errorMessage() {
+        return "Error ({$this->getFile()}:{$this->getLine()}) -> {$this->getMessage()}";
+    }
+}
+// ENVIRONMENT NOT FOUND
+class environmentNotFoundException extends Exception {
+    public function errorMessage() {
+        return "Error ({$this->getFile()}:{$this->getLine()}) -> {$this->getMessage()}";
+    }
+}
+// ENVIRONMENT VARIABLE NOT FOUND
+class environmentVariableNotFoundException extends Exception {
+    public function errorMessage() {
+        return "Error ({$this->getFile()}:{$this->getLine()}) -> {$this->getMessage()}";
+    }
+}
+
 // EXCEPTION HANDLER
 function exceptionHandler(Throwable $exception) {
     echo $exception->getMessage();