+++ website/modules/test.php
@@ -1,0 +1,7 @@
+<?php 
+function test($pattern, $value, $name) {
+    if (!preg_match($pattern, $value)) {
+        throw new regexException($name);
+    }
+}
+?>